requirements.txt
flask==2.0.1
whoosh==2.7.4
requests==2.26.0
googlemaps==4.6.0
python-dotenv==0.19.0
firebase-admin==5.0.0
gunicorn==21.2.0
